Make many things async and await Lisp#run
Lisp#run is an async function, so everything that calls it should also
be async. I've opted for explicitly making things async even when the JS
API that calls it doesn't do anything to specifically handle the promise
to help document expectations. This will make it easier to do spot
places that need error handling later.

This means _run only has "active" for as long as it's actually running.